chagned activate mthod accordingly to then ew interface Activations

+++ src/main/java/NN/Sigmoid.java
@@ -6,6 +6,6 @@ public class Sigmoid implements Activations<Double>{
 
     //method for the sigmoid function
     public Double activate(Double x){
-        return(new Double(1.0 / (1.0 + Math.exp(-x))));
+        return(new Double(1.0 / (1.0 + Math.exp(-x.doubleValue()))));
     }
 }
